+++ include/coreir/common/wiring_utils.h
