+++ include/coreir/simulator/layout.h
@@ -3,8 +3,8 @@
 #include "coreir/ir/value.h"
 #include "coreir/ir/wireable.h"
 
-#include "coreir/simulator/op_graph.h"
 #include "coreir/simulator/print_c.h"
+#include "coreir/common/op_graph.h"
 #include "coreir/common/utils.h"
 
 namespace CoreIR {
