+++ src/simulator/subcircuit.cpp
@@ -4,8 +4,8 @@
 #include "coreir/ir/types.h"
 #include "coreir/passes/transform/deletedeadinstances.h"
 #include "coreir/passes/transform/fold_constants.h"
-#include "coreir/simulator/op_graph.h"
-#include "coreir/simulator/wiring_utils.h"
+#include "coreir/common/op_graph.h"
+#include "coreir/common/wiring_utils.h"
 #include "coreir/common/algorithm.h"
 #include "coreir/common/utils.h"
